// SPES-client/src/model/pfInfoWithIdsForUpdate.ts
/**
 * SPES
 * ### Documentazione Termini chiave:   - PF: persona fragile   - MED: medico   - OPS: Operatore Sociosanitario (o chi ne fa le veci)   Autori:   - Luca Gregori   - Alessandro Wood
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MaritalStatus } from './maritalStatus';
import { Address } from './address';
import { Citizenship } from './citizenship';


export interface PfInfoWithIdsForUpdate {
    pf_id: string;
    cf?: string;
    firstname: string;
    lastname: string;
    fullname?: string;
    gender?: string;
    nicknames?: string;
    birth_date?: string;
    birth_nation_id?: string;
    birth_geoarea_id?: number;
    birth_city?: string;
    cui_code?: string;
    sanitary_district_id?: number;
    is_foreign?: boolean;
    is_anonymous?: boolean;
    verified?: boolean;
    is_dead?: boolean;
    death_date?: string;
    marital_status_list_to_add?: Array<MaritalStatus>;
    address_list_to_add?: Array<Address>;
    citizenship_list_to_add?: Array<Citizenship>;
    marital_status_list_to_delete?: Array<string>;
    address_list_to_delete?: Array<string>;
    citizenship_list_to_delete?: Array<string>;
    marital_status_list_to_update?: { [key: string]: MaritalStatus; };
    address_list_to_update?: { [key: string]: Address; };
    citizenship_list_to_update?: { [key: string]: Citizenship; };
}